pyproject.toml
[project]
authors = [
    {name = "Luuk Verweij", email = "luuk_verweij@msn.com"},
]
requires-python = ">=3.12.1,<4.0"
dependencies = [
    "click>=8.1.3",
    "basil-parser==0.4.3",
]
name = "Aaa"
version = "0.1.0"
description = "Stack-based languge"

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 88  # Default width of Black
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
select = [  # Full list: https://docs.astral.sh/ruff/rules/
    "E",   # pycodestyle
    "F",   # pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]
ignore = [
    "SIM108", # use ternary instead of if-else
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "mypy>=1.7.1",
    "pdbpp>=0.10.3",
    "pre-commit>=3.0.1",
    "pytest>=7.2.1",
    "pytest-cov>=4.0.0",
    "types-click>=7.1.8",
    "requests>=2.31.0",
    "types-requests>=2.31.0.1",
    "ruff>=0.1.8",
]

[tool.pdm.build]
includes = []
